+++ productionized/pulse-fetch/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Fixed flaky `strategy-config-env` functional test caused by race condition with `default-config` test over shared `/tmp/pulse-fetch` directory
+
 ## [0.3.0] - 2025-10-08
 
 ### Changed